+++ Properties/launchSettings.json
@@ -1,0 +1,2 @@
+{
+}
